HACK: wahoo: backend: Mount partitions after the device check
getprop in TWRP on the Pixel 2 XL on P DP4 currently fails to get the
props with this section in its current place. Mounting the partitions
after getting the prop values fixes this.

recovery.log in TWRP 3.2.2-0 without this commit:
* OnePlus 6 (OxygenOS 5.1.9): https://gist.github.com/6b814dd3d6c1d4c714ca07737b3bc83a
* Pixel XL (P DP4): https://gist.github.com/379f55f1cc28bb9147ca983971974247
* Pixel 2 XL (Oreo, July 2018): https://gist.github.com/01bb7cc5fcc3784b3984df4365e50f94
* Pixel 2 XL (P DP4): https://gist.github.com/81ef6ae351331db847d9f41c22f20f7d

recovery.log in TWRP 3.2.2-0 with this commit:
* OnePlus 6 (OxygenOS 5.1.9): https://gist.github.com/e2d4b1f248aee24bcc4bfe857fb29b57
* Pixel XL (P DP4): https://gist.github.com/a0097699b2adee01c034f9e4d3809e86
* Pixel 2 XL (Oreo, July 2018): https://gist.github.com/6d96cf69a6e0ee57031d5907ceb2356e
* Pixel 2 XL (P DP4): https://gist.github.com/62ee4aa8c8a09392c88253585d18fa9c

TL;DR: Pixel 2 XL P DP4 is the only one that is broken without the
commit but everything works after it.

HOWEVER, as this is backend code, it should NOT be cherry-picked by
anyone else to minimize conflicts with upstream. A proper solution is in
the works but users shouldn't be barred from flashing this kernel in the
meantime.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>